# GH-41541: [Go][Parquet] Fix writer performance regression