@zeroshade 2024-05-13
@zhouyan can you give this a test to confirm that it fixes the performance degradation you saw?