@ggodik 2024-05-13
successfully confirmed this restored write performance to pre v14 levels

thanks @zeroshade